Changed structure of directories and files linked to videochat funcionality
