+++ src/index.html
@@ -23,7 +23,7 @@
         crossorigin="anonymous"></script>
 
     <script src="https://cdn.jsdelivr.net/npm/peerjs@0.3.20/dist/peer.min.js"></script>
-    <script src="lib/VideoChat/PeerManager.js"></script>
+    <script src="lib/videochatManager.js"></script>
